+++ MyGameProject/Game/PlayerJumpThrowState.h
@@ -10,7 +10,7 @@ public:
 	void Render() override;
 	void Update(float dt) override;
 	void HandleInput() override;
-	void OnCollision(Entity* impactor, Entity::SideCollision side, float collisionTime, double dt = 1.0 / 60) override;
+	void OnCollision(Entity* impactor, Entity::SideCollision side, float collisionTime, float dt = 1.0 / 60) override;
 	State GetStateName() override;
 	void ResetState(int dummy) override;
 };
